ログイン処理をuseLogin.jsxにまとめて使う形にした
